move Message component to separate file
