+++ src/components/MessageList/MessageList.tsx
@@ -4,6 +4,8 @@ import PropTypes from 'prop-types'
 import cx from 'classnames'
 import { createComponent, customPropTypes } from '../../lib'
 
+import Message from './Message'
+
 const messageListRules = () => ({
   root: {
     listStyle: 'none',
@@ -12,45 +14,6 @@ const messageListRules = () => ({
     margin: 0,
   },
 })
-
-const messageRules = ({ isMine, variables }) => ({
-  container: {
-    display: 'flex',
-    margin: '1rem', // TODO is temporary hardcoded value
-    justifyContent: (isMine && 'flex-end') || 'flex-start',
-  },
-  text: {
-    display: 'inline-block',
-    backgroundColor: (isMine && variables.messageColor.mine) || variables.messageColor.incoming,
-    maxWidth: variables.messageWidth,
-
-    // TODO - the following ones are temporay
-    padding: '1rem',
-    borderRadius: '0.3rem',
-    color: 'rgb(64,64,64)',
-    fontFamily: 'Segoe UI, sans-serif',
-  },
-})
-
-const messageVariables = {
-  messageWidth: '80%',
-  messageColor: {
-    mine: '#E0E0ED',
-    incoming: 'rgba(0,0,0,0.1)',
-  },
-}
-
-const Message = createComponent(
-  ({ content, styles, className }) => (
-    <div className={cx('ui-message__container', styles.container, className)}>
-      <p className={cx('ui-message__text', styles.text)}>{content}</p>
-    </div>
-  ),
-  {
-    rules: messageRules,
-    variables: messageVariables,
-  },
-)
 
 const MessageList: any = props => (
   <ul className={cx('ui-message__list', props.styles.root, props.className)}>
@@ -69,6 +32,8 @@ MessageList.propTypes = {
 
   /** Shorthand array of message items. */
   items: PropTypes.arrayOf(PropTypes.any),
+
+  children: PropTypes.node,
 }
 
 // TODO introduce
